refactor: [BE] Order 객체를 생성하는 복잡한 로직을 Order Entity에서 of 함수 구현
